requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.2
flask-cors==3.0.10
gunicorn==20.1.0